Decrypt applet responses after download
